% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/re.R
\name{re}
\alias{re}
\title{Random effect parser}
\usage{
re(variable)
}
\arguments{
\item{variable}{id used to identify clusters}
}
\description{
Internal function used to parse random effects in the formula interface
to other functions.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

