% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VB_Mult_Wish.R
\name{vb_mult_wish}
\alias{vb_mult_wish}
\title{Multilevel FoSR using Variational Bayes and Wishart prior}
\usage{
vb_mult_wish(formula, data = NULL, Kt = 5, alpha = 0.1)
}
\arguments{
\item{formula}{a formula indicating the structure of the proposed model.}

\item{data}{an optional data frame, list or environment containing the
variables in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which the function is
called.}

\item{Kt}{number of spline basis functions used to estimate coefficient functions}

\item{alpha}{tuning parameter balancing second-derivative penalty and
zeroth-derivative penalty (alpha = 0 is all second-derivative penalty)}
}
\description{
Fitting function for function-on-scalar regression for cross-sectional data.
This function estimates model parameters using VB and estimates
the residual covariance surface using a Wishart prior.
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}
\references{
Goldsmith, J., Kitago, T. (Under Review).
Assessing Systematic Effects of Stroke on Motor Control using Hierarchical
Function-on-Scalar Regression.
}

